refactored fgo_nn_splines
