+++ experiments/fgo_nn_splines.py
@@ -29,148 +29,40 @@ from ronin_resnet import get_model
 from ronin_resnet import ResNet1D, BasicBlock1D, FCOutputModule
 from model_temporal import TCNSeqNetwork
 
-def compute_delta_x(gt_poses, estimated_poses):
-    gt_poses_se3 = [mrob.SE3(gt_poses[i]) for i in range(len(gt_poses))]
-    estimated_poses_se3 = [mrob.SE3(estimated_poses[j]) for j in range(len(estimated_poses))]
-    
-    result_Ln = [(gt_poses_se3[k] * estimated_poses_se3[k].inv()).Ln() for k in range(len(gt_poses_se3))]
-    return np.array(result_Ln)
+from pathlib import Path
+import torch
+import torch.optim as optim
+import torch.nn as nn
+from spline_dataset.spline_dataloader import Spline_2D_Dataset
+from torch.utils.data import DataLoader
+from tqdm import tqdm
+from torch.multiprocessing import Pool
+import numpy as np
+np.set_printoptions(precision=4,linewidth=180)
+import pickle
+from torch.utils.tensorboard import SummaryWriter
+import matplotlib.pyplot as plt
 
-
-def integrate_pred_vel(pred_vel, gt_poses, gt_pose0, dt=0.01):
-    T = pred_vel.shape[0]
-    poses = torch.zeros((T, 3), dtype=pred_vel.dtype, device=pred_vel.device)
-
-    x, y, yaw = gt_pose0[0], gt_pose0[1], gt_pose0[2]
-    poses[0] = torch.stack([x, y, yaw])
-
-    for t in range(1, T):
-        vx, vy = pred_vel[t - 1]
-
-        x = x + vx * dt
-        y = y + vy * dt
-        yaw = gt_poses[t, 2]  # directly use GT yaw
-
-        poses[t] = torch.stack([x, y, yaw])
-
-    return poses
-
-
-def populate_graph(pred_poses, gt_poses):
-    graph = mrob.FGraphDiff()
-
-    W_odo = torch.eye(6) * 10.0
-    W_gps = torch.eye(6) * 1.0
-
-    node_ids = []
-    T_nodes = []
-    T = pred_poses.shape[0]
-    
-    # Initialize nodes using predicted poses (world frame)
-    for i in range(T): # неподвижная система координат
-        pose_i = pred_poses[i].detach().cpu().numpy()
-        T_i = mrob.SE3(convert_to_se3(pose_i))
-        T_nodes.append(T_i)
-        node_id = graph.add_node_pose_3d(T_i)
-        node_ids.append(node_id)
-
-    # Odometry factors from predicted poses
-    for i in range(len(node_ids) - 1):
-        T_pred_i   = T_nodes[i]
-        T_pred_ip1 = T_nodes[i + 1]
-        T_odo = T_pred_i.inv() * T_pred_ip1 #mrob.SE3(pred_v, w (from IMU))
-        graph.add_factor_2poses_3d_diff(T_odo, node_ids[i], node_ids[i + 1], W_odo.numpy())
-
-    # GPS factors from GT poses
-    for i in range(len(node_ids)):
-        # if i % 10 == 0 or i == len(node_ids) - 1:
-        T_gps = mrob.SE3(convert_to_se3(gt_poses[i].detach().cpu().numpy()))
-        graph.add_factor_1pose_3d_diff(T_gps, node_ids[i], W_gps.numpy())
-    
-    return graph
-
-
-def print_2d_graph(graph, gt_poses):
-    x = np.array(graph.get_estimated_state())
-    
-    fig = plt.figure()
-
-    plt.plot(x[:,0,3],x[:,1,3], '-b', marker='o', label='estimated')
-    plt.plot(gt_poses[:, :2][:, 0], gt_poses[:, :2][:, 1], label='GT', color='red')
-    plt.title("2D Pose Graph")
-
-    plt.xlabel("X")
-    plt.ylabel("Y")
-    plt.legend()
-    plt.axis('equal')
-    plt.grid()
-    # plt.show()
-    return fig
-
-
-def make_gif_from_figures(folder_path, output_path="graph_evolution.gif", num_epochs=100, duration=0.3):
-    file_list = sorted(
-        [f for f in os.listdir(folder_path) if f.startswith("poses_") and f.endswith(".png") ],
-        key=lambda x: int(''.join(filter(str.isdigit, x)))
-    )
-    
-    file_list = file_list[0:num_epochs]
-
-    images = [imageio.imread(os.path.join(folder_path, fname)) for fname in file_list]
-    imageio.mimsave(output_path, images, duration=duration)
-    print(f"GIF saved to {output_path}")
-
-
-def plot_integrated_vel(poses_1, poses_2):
-    plt.figure()
-    plt.plot(poses_1[:, 0].detach(), poses_1[:, 1].detach(), marker='o', label='integrated')
-    plt.plot(poses_2[:, 0].detach(), poses_2[:, 1].detach(), marker='o', label='gt')
-    plt.legend()
-    plt.show()
-    return
-
-def process_one_graph(vel_pred, gt_pose_seq, dt):
-    S = vel_pred.shape[0]
-    # Step 1: integrate vel_pred[b] into poses
-    pred_poses = integrate_pred_vel(vel_pred, gt_pose_seq, gt_pose_seq[0], dt=dt)  # [S, 3]
-    
-    # Step 2: build & solve graph
-    graph = populate_graph(pred_poses, gt_pose_seq)  # FGraphDiff
-    graph.build_jacobians()
-    graph.solve(mrob.FGraphDiff_LM, maxIters=100)
-    dL_dz = graph.get_dx_dz() / dt  # [6S, 6S + (S-1)*S]
-
-    # Step 3: compute delta_x
-    gt_poses_se3 = convert_to_se3(gt_pose_seq.cpu().numpy())
-    delta_x = compute_delta_x(gt_poses_se3, graph.get_estimated_state())  # [S, 6]
-    delta_x_flat = delta_x.reshape(-1)  # [6S]
-
-    # Step 4: compute gradient per sample and stack
-    grad = (delta_x_flat @ dL_dz)[:S*6].reshape(-1, 6)
-    grad_final = torch.from_numpy(grad[:, [3, 4]]).float()  # [S, 2]
-
-    # Step 5: computing chi2 and rmse errors
-    chi2 = graph.chi2()
-    rmse = compute_rmse_and_yaw(graph, gt_poses_se3, delta_x, plot=False)
-    
-    return grad_final, chi2, rmse
+from ronin_resnet import get_model
+from ronin_resnet import ResNet1D, BasicBlock1D, FCOutputModule
+from fgo.fgo_implementation import process_one_graph, integrate_pred_vel, populate_graph
 
 def run_validation(epoch, output_path, model, dataset, num_traj, device):
     model.eval()
 
     assert num_traj <= len(dataset)
 
-    if not os.path.exists(output_path):
-        os.makedirs(output_path,exist_ok=True)
 
     for i in range(num_traj):
         result = {}
 
         sample  = dataset.__getitem__(i)
         imu_seq = sample['noisy_imu'].to(device)
-        vel_seq = sample['gt_vel']
         gt_poses_seq = sample['gt_poses']
         dt = dataset.step/dataset.sampling_rate
+
+        trajectories_path = output_path / "trajectories"
+        trajectories_path.mkdir(parents=True, exist_ok=True)
 
         S, C, W = imu_seq.shape
         vel_pred = model(imu_seq.reshape(-1, C, W)).reshape(S, -1).detach().cpu()
@@ -205,13 +97,12 @@ def run_validation(epoch, output_path, model, dataset, num_traj, device):
         plt.title(f'Epoch: {epoch}\n' + \
             f"ATE: {ate:.3f}, RTE: {rte:.3f}")
         plt.tight_layout()
-        plt.savefig(f'{output_path}idx_{i}_epoch_{epoch}.jpg')
+        plt.savefig(trajectories_path / f'idx_{i}_epoch_{epoch}.jpg')
         plt.close('all')
-        pickle.dump(result, open(f'{output_path}idx_{i}_epoch_{epoch}_traj.pkl','wb'))
+        pickle.dump(result, open(output_path / f'idx_{i}_epoch_{epoch}_traj.pkl','wb'))
 
-
-
-def run_spline_experiment(subseq_len = 3, n_epochs=300):
+def fgo_nn_splines_train_loop(train_dataloader, val_dataloader=None, subseq_len = 3, 
+                          n_epochs=300, output_path=None, device="cpu", start_lr=1e-3,):
     '''
     Odometry model training pipeline on spline dataset
     if subseq_len == 1 - conventional window-based training mode
@@ -223,10 +114,11 @@ def run_spline_experiment(subseq_len = 3, n_epochs=300):
     results['n_actual_epochs'] = 0
     results['subseq_len'] = subseq_len
 
-    output_path = f"./out/graphs_seq_{subseq_len}_epochs_{n_epochs}/"
-    if not os.path.exists(output_path):
-        os.makedirs(output_path, exist_ok=True)
-    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+    output_path = Path(output_path)
+    output_path.mkdir(parents=True, exist_ok=True)
+
+    device = torch.device(device)
+
     model = ResNet1D(
         num_inputs=3,       
         num_outputs=2,         
@@ -241,38 +133,13 @@ def run_spline_experiment(subseq_len = 3, n_epochs=300):
         trans_planes=128
     ).to(device)
 
-    # model.load_state_dict(torch.load("out/model_fgo.pth"))
-    optimizer = optim.Adam(model.parameters(), lr=1e-3)
+    optimizer = optim.Adam(model.parameters(), lr=start_lr)
     scheduler = torch.optim.lr_scheduler.StepLR(optimizer, step_size=100, gamma=0.9)
 
     criterion = nn.MSELoss()
 
-    path_to_splines =  './out/splines'
-
-    number_of_splines = 20
-    if not os.path.exists(path_to_splines):
-        number_of_control_nodes = 10
-        generate_batch_of_splines(path_to_splines, number_of_splines, number_of_control_nodes, 100)
-        
-    window_size=100
     step_size=10
-    sampling_rate =100
-    
-    dataset = Spline_2D_Dataset(path_to_splines, 
-                                window=window_size,
-                                sampling_rate=100,
-                                subseq_len=subseq_len,
-                                mode='regression',
-                                enable_noise= not True)
-
-    train_dataloader = DataLoader(dataset, batch_size=64, shuffle=True, collate_fn=dataset.get_collate_fn())
-
-    val_dataset = Spline_2D_Dataset(path_to_splines, window=window_size, subseq_len=89, enable_noise= not True, is_val = True)
-    val_dataloader = DataLoader(val_dataset, batch_size=1, shuffle=False)
-
-    save_file_split(path_to_splines, output_path)
-    
-    dt = step_size/sampling_rate
+    dt = step_size/train_dataloader.dataset.sampling_rate
     rmse_errors = []
     chi2_errors = []
     learning_rates = []
@@ -280,15 +147,17 @@ def run_spline_experiment(subseq_len = 3, n_epochs=300):
     trajectories_to_save = 3
     results['num_val_traj'] = trajectories_to_save
 
+
+    writer = SummaryWriter(output_path)
+    
     for epoch in range(n_epochs):
 
         # running validation every epoch
-        run_validation(epoch, output_path, model, val_dataloader.dataset, trajectories_to_save, device)
+        if val_dataloader:
+            run_validation(epoch, output_path, model, val_dataloader.dataset, trajectories_to_save, device)
         total_chi2, total_rmse = 0.0, 0.0
         model.train()
-        
         for sample in tqdm(train_dataloader, position=0, leave=True):
-            # sample  = dataset.__getitem__(i)
             imu_seq = sample['noisy_imu'].to(device)
             vel_seq = sample['gt_vel'].to(device)
             gt_poses_seq = sample['gt_poses']
@@ -303,7 +172,6 @@ def run_spline_experiment(subseq_len = 3, n_epochs=300):
             vel_pred = model(imu_seq.reshape(-1, C, W)).reshape(B, S, -1)
             
             optimizer.zero_grad()
-
 
             if subseq_len > 1:
                 all_grads = [None for _ in range(B)]
@@ -322,9 +190,8 @@ def run_spline_experiment(subseq_len = 3, n_epochs=300):
 
                 grad_tensor = torch.stack(all_grads).to(device)  # [B, S, 2]
                 vel_pred.backward(gradient= - grad_tensor)
-
             else:
-                loss = torch.linalg.norm(vel_pred - vel_seq)
+                loss = criterion(vel_pred, vel_seq)
                 loss.backward()
 
                 total_rmse += loss.detach().cpu().item()
@@ -333,7 +200,7 @@ def run_spline_experiment(subseq_len = 3, n_epochs=300):
             torch.nn.utils.clip_grad_norm_(model.parameters(), max_norm=0.1)
             optimizer.step()
         scheduler.step()
-        
+
         chi2_errors.append(total_chi2 / len(train_dataloader))
         rmse_errors.append(total_rmse / len(train_dataloader))
         learning_rates.append(scheduler.get_last_lr()[0])
@@ -345,32 +212,46 @@ def run_spline_experiment(subseq_len = 3, n_epochs=300):
         results['chi2_errors'] = chi2_errors
         results['rmse_errors'] = rmse_errors
         results['learning_rate'] = learning_rates
-        pickle.dump(results, open(output_path+'results.pkl','wb'))
+
+        writer.add_scalar("chi2_errors", chi2_errors[-1], results['n_actual_epochs'])
+        writer.add_scalar("rmse_errors", rmse_errors[-1], results['n_actual_epochs'])
+        writer.add_scalar("learning_rates", learning_rates[-1], results['n_actual_epochs'])
+
+        pickle.dump(results, open(output_path / "trajectories" / 'results.pkl','wb'))
 
         if epoch % 10 == 0:
-            torch.save(model, output_path + f'model_epoch_{epoch}.cpt')
+            torch.save(model, output_path / f'model_epoch_{epoch}.cpt')
 
-    plt.title('Errors: CHi2 and RMSE')
-    plt.plot(chi2_errors,label='chi2')
-    plt.plot(rmse_errors,label='rmse')
-    plt.xlabel('epoch')
-    plt.grid()
-    plt.legend()
-    plt.savefig(f'{output_path}/errors.png')
-    plt.close('all')
 
-    plt.figure()
-    plt.plot(learning_rates,label='learning rate')
-    plt.grid()
-    plt.xlabel('epoch')
-    plt.legend()
-    plt.savefig(f'{output_path}/learning_rate.png')
-    plt.close('all')
+
+
+
 
 if __name__ == "__main__":
+    subseq_len=2
+    n_epochs=31
+    output_path = f"./out/graphs_seq_{subseq_len}_epochs_{n_epochs}_testing_subseq_length/"
+    path_to_splines = "./out/splines"
+    window_size=100
 
-    # for s in range(1,5,1):
-    #     run_spline_experiment(s, 50)
+    output_path = Path(output_path)
+    output_path.mkdir(parents=True, exist_ok=True)
 
-    run_spline_experiment(1, 100)
-    run_spline_experiment(6, 100)
+    train_dataset = Spline_2D_Dataset(path_to_splines, 
+                                window=window_size,
+                                sampling_rate=100,
+                                subseq_len=subseq_len,
+                                mode='regression',
+                                enable_noise= not True)
+
+    train_dataloader = DataLoader(train_dataset, batch_size=64, shuffle=True, collate_fn=train_dataset.get_collate_fn())
+
+    # There was subseq_len equal to 90. Why?
+    val_dataset = Spline_2D_Dataset(path_to_splines, window=window_size, subseq_len=subseq_len, enable_noise= not True, is_val = True)
+    val_dataloader = DataLoader(val_dataset, batch_size=1, shuffle=False, collate_fn=val_dataset.get_collate_fn())
+
+    save_file_split(path_to_splines, output_path)
+
+    device = torch.device('cuda:0' if torch.cuda.is_available()  else 'cpu')
+    fgo_nn_splines_train_loop(train_dataloader, val_dataloader=val_dataloader, subseq_len = subseq_len, 
+                          n_epochs=n_epochs, output_path=output_path, device=device, start_lr=1e-3)
